+++ NEWS.md
@@ -1,3 +1,10 @@
+# phstemplate 1.0.0
+
+* Added the phs_report_docx RMarkdown document type.
+* Updated the PHS official and national stats report templates to use the phs_report_docx RMarkdown document type.
+* Removed the compile_report() function.
+* Removed the ISD national stats summary and report templates.
+
 # phstemplates 0.9.6
 
 * Added PHS official and national stats summary and report documents for 2022.